Add files via upload
1. Identify the Target 
Machine's IP Address
Use the netdiscover command within the Kali Linux VM to identify the IP address of the target machine on the same network.
2. Determine Open Ports and Services
Execute the Nmap tool to scan the target machine's IP address and list open ports and running services. Nmap is available by default in Kali Linux.
3. Identify Vulnerabilities in Web Application
Based on the Nmap scan results, access the web application hosted on the target machine. Attempt to log in using default or discovered credentials.
4. Directory Enumeration Using Dirb
Utilize the dirb utility to perform directory enumeration on the web application. Search for hidden directories or files that may contain sensitive 
information.
5. Login Using Detected Credentials
Use the username and password obtained during the previous steps to log in to  the web application. Document this process with relevant screenshots